liquidite management more reactive
