+++ Frontend-v1-Original/stores/constants/actions.ts
@@ -47,11 +47,11 @@ export const WARNING = "WARNING";
 // export const ADD_LIQUIDITY = "ADD_LIQUIDITY";
 // export const LIQUIDITY_ADDED = "LIQUIDITY_ADDED";
 
-export const STAKE_LIQUIDITY = "STAKE_LIQUIDITY";
-export const LIQUIDITY_STAKED = "LIQUIDITY_STAKED";
+// export const STAKE_LIQUIDITY = "STAKE_LIQUIDITY";
+// export const LIQUIDITY_STAKED = "LIQUIDITY_STAKED";
 
-export const ADD_LIQUIDITY_AND_STAKE = "ADD_LIQUIDITY_AND_STAKE";
-export const ADD_LIQUIDITY_AND_STAKED = "ADD_LIQUIDITY_AND_STAKED";
+// export const ADD_LIQUIDITY_AND_STAKE = "ADD_LIQUIDITY_AND_STAKE";
+// export const ADD_LIQUIDITY_AND_STAKED = "ADD_LIQUIDITY_AND_STAKED";
 
 // export const QUOTE_ADD_LIQUIDITY = "QUOTE_ADD_LIQUIDITY";
 // export const QUOTE_ADD_LIQUIDITY_RETURNED = "QUOTE_ADD_LIQUIDITY_RETURNED";
@@ -73,8 +73,8 @@ export const REMOVE_LIQUIDITY_AND_UNSTAKED = "REMOVE_LIQUIDITY_AND_UNSTAKED";
 // export const QUOTE_REMOVE_LIQUIDITY_RETURNED =
 //   "QUOTE_REMOVE_LIQUIDITY_RETURNED";
 
-export const CREATE_GAUGE = "CREATE_GAUGE";
-export const CREATE_GAUGE_RETURNED = "CREATE_GAUGE_RETURNED";
+// export const CREATE_GAUGE = "CREATE_GAUGE";
+// export const CREATE_GAUGE_RETURNED = "CREATE_GAUGE_RETURNED";
 
 // //SWAPS
 // export const SWAP = "SWAP";
@@ -113,8 +113,8 @@ export const MERGE_NFT_RETURNED = "MERGE_NFT_RETURNED";
 // export const VOTE = "VOTE";
 // export const VOTE_RETURNED = "VOTE_RETURNED";
 
-export const GET_VEST_VOTES = "GET_VEST_VOTES";
-export const VEST_VOTES_RETURNED = "VEST_VOTES_RETURNED";
+// export const GET_VEST_VOTES = "GET_VEST_VOTES";
+// export const VEST_VOTES_RETURNED = "VEST_VOTES_RETURNED";
 
 // export const CREATE_BRIBE = "CREATE_BRIBE";
 // export const BRIBE_CREATED = "BRIBE_CREATED";
